third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/preload/onerror_event-expected.txt

CONSOLE WARNING: line 26: <link rel=preload> must have a valid `as` value
This is a testharness.js-based test.
PASS Makes sure that preloaded resources trigger the onerror event 
Harness: the test ran to completion.
